# BookStore.RestApi.Client/Api/swagger.json
{
  "openapi": "3.0.1",
  "info": {
    "title": "BookStore.WepApi.Host",
    "version": "1.0"
  },
  "paths": {
    "/api/Author/book/{bookId}": {
      "get": {
        "tags": [
          "Author"
        ],
        "summary": "Получение списка авторов издания",
        "parameters": [
          {
            "name": "bookId",
            "in": "path",
            "description": "Идентификатор издания",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/AuthorDto"
                  }
                }
              },
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/AuthorDto"
                  }
                }
              },
              "text/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/AuthorDto"
                  }
                }
              }
            }
          },
          "204": {
            "description": "No Content"
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      }
    },
...
    "/api/Book/{id}": {
      "put": {
        "tags": [
          "Book"
        ],
        "summary": "Изменение имеющихся данных",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "Идентификатор",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "requestBody": {
          "description": "Измененные данные",
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/BookCreateUpdateDto"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/BookCreateUpdateDto"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/BookCreateUpdateDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/BookDto"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BookDto"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/BookDto"
                }
              }
            }
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      },
      "delete": {
        "tags": [
          "Book"
        ],
        "summary": "Удаление данных",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "Идентификатор",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "204": {
            "description": "No Content"
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      },
      "get": {
        "tags": [
          "Book"
        ],
        "summary": "Получение данных по идентификатору",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "Идентификатор",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/BookDto"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BookDto"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/BookDto"
                }
              }
            }
          },
          "204": {
            "description": "No Content"
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "AuthorDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "description": "Идентификатор",
            "format": "int32"
          },
          "lastName": {
            "type": "string",
            "description": "Фамилия",
            "nullable": true
          },
          "firstName": {
            "type": "string",
            "description": "Имя",
            "nullable": true
          },
          "patronymic": {
            "type": "string",
            "description": "Отчество",
            "nullable": true
          },
          "biography": {
            "type": "string",
            "description": "Биография",
            "nullable": true
          },
          "workCount": {
            "type": "integer",
            "description": "Число работ",
            "format": "int32",
            "nullable": true
          }
        },
        "additionalProperties": false,
        "description": "Dto для просмотра сведений об авторе"
      },
...
      "BookCreateUpdateDto": {
        "type": "object",
        "properties": {
          "title": {
            "type": "string",
            "description": "Название",
            "nullable": true
          },
          "annotation": {
            "type": "string",
            "description": "Аннотация",
            "nullable": true
          },
          "pageCount": {
            "type": "integer",
            "description": "Число страниц",
            "format": "int32",
            "nullable": true
          },
          "year": {
            "type": "integer",
            "description": "Год",
            "format": "int32",
            "nullable": true
          },
          "publisher": {
            "type": "string",
            "description": "Издательство",
            "nullable": true
          },
          "isbn": {
            "type": "string",
            "description": "ISBN",
            "nullable": true
          }
        },
        "additionalProperties": false,
        "description": "Dto для создания или изменения издания"
      },
      "BookDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "description": "Идентификатор",
            "format": "int32"
          },
          "title": {
            "type": "string",
            "description": "Название",
            "nullable": true
          },
          "annotation": {
            "type": "string",
            "description": "Аннотация",
            "nullable": true
          },
          "pageCount": {
            "type": "integer",
            "description": "Число страниц",
            "format": "int32",
            "nullable": true
          },
          "year": {
            "type": "integer",
            "description": "Год",
            "format": "int32",
            "nullable": true
          },
          "publisher": {
            "type": "string",
            "description": "Издательство",
            "nullable": true
          },
          "isbn": {
            "type": "string",
            "description": "ISBN",
            "nullable": true
          }
        },
        "additionalProperties": false,
        "description": "Dto для просмотра сведений об издании"
      }
    }
  }
}